Moved auth to helpers
