+++ token/token.go
@@ -1,8 +1,6 @@
 package token
 
 import (
-	"crypto/sha256"
-	"crypto/subtle"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -18,29 +16,6 @@ var Password = "uppmax"
 type tokenRequest struct {
 	Swamid    string `json:"swamid"`
 	ProjectID string `json:"projectid"`
-}
-
-func BasicAuth(next http.HandlerFunc) http.HandlerFunc {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		username, password, ok := r.BasicAuth()
-		if ok {
-			usernameHash := sha256.Sum256([]byte(username))
-			passwordHash := sha256.Sum256([]byte(password))
-			expectedUsernameHash := sha256.Sum256([]byte(Username))
-			expectedPasswordHash := sha256.Sum256([]byte(Password))
-
-			usernameMatch := (subtle.ConstantTimeCompare(usernameHash[:], expectedUsernameHash[:]) == 1)
-			passwordMatch := (subtle.ConstantTimeCompare(passwordHash[:], expectedPasswordHash[:]) == 1)
-
-			if usernameMatch && passwordMatch {
-				next.ServeHTTP(w, r)
-				return
-			}
-		}
-
-		w.Header().Set("WWW-Authenticate", `Basic realm="restricted", charset="UTF-8"`)
-		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-	})
 }
 
 func readRequestBody(body io.ReadCloser) (tokenRequest tokenRequest, err error) {
@@ -65,7 +40,6 @@ func readRequestBody(body io.ReadCloser) (tokenRequest tokenRequest, err error) 
 
 func GetToken(w http.ResponseWriter, r *http.Request) {
 
-	//tokenRequest := tokenRequest{}
 	tokenRequest, err := readRequestBody(r.Body)
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	if err != nil {
